refactor transform api input wrappers
